Move unused declaration support into its own package
